requirements/test.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    make upgrade
#
-e git+https://github.com/edx/codejail.git@3.0.1#egg=codejail==3.0.1  # via -r requirements/base.txt
appdirs==1.4.4            # via -r requirements/base.txt, fs
attrs==20.2.0             # via pytest
bleach==3.2.1             # via -r requirements/base.txt
certifi==2020.6.20        # via -r requirements/base.txt, requests
chardet==3.0.4            # via -r requirements/base.txt, requests
chem==1.2.0               # via -r requirements/base.txt
click==7.1.2              # via -r requirements/base.txt, code-annotations, nltk
code-annotations==0.10.0  # via -r requirements/test.in
coverage==5.3             # via pytest-cov
ddt==1.4.1                # via -r requirements/test.in
fs==2.4.11                # via -r requirements/base.txt
html5lib==1.1             # via -r requirements/base.txt
idna==2.10                # via -r requirements/base.txt, requests
iniconfig==1.1.1          # via pytest
jinja2==2.11.2            # via code-annotations
joblib==0.17.0            # via -r requirements/base.txt, nltk
lxml==4.6.1               # via -r requirements/base.txt
mako==1.1.3               # via -r requirements/base.txt
markupsafe==1.1.1         # via -r requirements/base.txt, chem, jinja2, mako
mpmath==1.1.0             # via -r requirements/base.txt, sympy
nltk==3.5                 # via -r requirements/base.txt, chem
numpy==1.19.4             # via -r requirements/base.txt, chem, openedx-calc, scipy
openedx-calc==1.0.9       # via -r requirements/base.txt
packaging==20.4           # via -r requirements/base.txt, bleach, pytest
path==15.0.0              # via -r requirements/base.txt
pbr==5.5.1                # via stevedore
pluggy==0.13.1            # via pytest
py==1.9.0                 # via pytest
pyparsing==2.4.7          # via -r requirements/base.txt, chem, openedx-calc, packaging
pytest-cov==2.10.1        # via -r requirements/test.in
pytest-django==4.1.0      # via -r requirements/test.in
pytest==6.1.2             # via pytest-cov, pytest-django
python-slugify==4.0.1     # via code-annotations
pytz==2020.4              # via -r requirements/base.txt, django, fs
pyyaml==5.3.1             # via code-annotations
random2==1.0.1            # via -r requirements/base.txt
regex==2020.10.28         # via -r requirements/base.txt, nltk
requests==2.24.0          # via -r requirements/base.txt
scipy==1.5.3              # via -r requirements/base.txt, chem, openedx-calc
shapely==1.7.1            # via -r requirements/base.txt
six==1.15.0               # via -r requirements/base.txt, bleach, chem, fs, html5lib, openedx-calc, packaging
sqlparse==0.4.1           # via -r requirements/base.txt, django
stevedore==3.2.2          # via code-annotations
sympy==1.6.2              # via -r requirements/base.txt
text-unidecode==1.3       # via python-slugify
toml==0.10.2              # via pytest
tqdm==4.51.0              # via -r requirements/base.txt, nltk
urllib3==1.25.11          # via -r requirements/base.txt, requests
webencodings==0.5.1       # via -r requirements/base.txt, bleach, html5lib

# The following packages are considered to be unsafe in a requirements file:
# setuptools
